[Fix] API関連の階層構造を修正
