+++ frontend/projects/src/app/api/movie/index.ts
@@ -1,0 +1,1 @@
+export * from "./createMovie";
